Create 01 DMA Report Summary.sql
After running an Assessment with the <abbr title="Data Migration Assistant">DMA</abbr> tool, export the report to a *.json file. Then run this script. It parses the JSON data to produce a Report Summary.

Here's a quick article for getting started with DMA: https://itsalljustelectrons.blogspot.com/2019/04/Data-Migration-Assistant-for-SQL-Server-Upgrades.html